feat(messages): open chat on new page
